# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: Centre Commercial API
  description: |
    API de gestion de centre commercial (MEAN stack).

    ## Rôles
    - **ADMIN** : gestion complète (boutiques, catégories, utilisateurs, stats)
    - **SHOP_MANAGER** : gestion de sa propre boutique (produits, promotions)
    - **CUSTOMER** : navigation, favoris, avis

    ## Authentification
    Stratégie double token JWT :
    - **Access token** : 15 min — à envoyer dans `Authorization: Bearer <token>`
    - **Refresh token** : 7 jours — à envoyer dans le body pour rafraîchir

    Les tokens ne sont délivrés qu'après vérification de l'email.
  version: '1.0.0'
servers:
  - url: http://localhost:3001/api
    description: Serveur de développement
security:
  - BearerAuth: []
...
paths:
...
  /admin/shops/{id}/approve:
    put:
      tags: [Admin — Shops]
      operationId: approveShop
      summary: Approuver une boutique (→ ACTIVE, manager → VALID)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Boutique approuvée
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Shop'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/shops/{id}/suspend:
    put:
      tags: [Admin — Shops]
      operationId: suspendShop
      summary: Suspendre une boutique (→ SUSPENDED)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Boutique suspendue
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Shop'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # ADMIN — CATEGORIES
  # ---------------------------------------------------------------------------
...
  /articles/create:
    post:
      tags: [Articles]
      operationId: createArticle
      summary: Créer un article (SHOP_MANAGER)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleInput'
      responses:
        '201':
          description: Article créé
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Article'
  /articles/update/{id}:
    put:
      tags: [Articles]
      operationId: updateArticle
      summary: Modifier un article (SHOP_MANAGER)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleInput'
      responses:
        '200':
          description: Article mis à jour
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Article'
        '404':
          $ref: '#/components/responses/NotFound'
  /articles/delete/{id}:
    delete:
      tags: [Articles]
      operationId: deleteArticle
      summary: Supprimer un article (SHOP_MANAGER)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Article supprimé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '404':
          $ref: '#/components/responses/NotFound'

# ===========================================================================
# COMPONENTS
# ===========================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: ID MongoDB de la ressource
...
  responses:
...
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
...
    UserRole:
      type: string
      enum: [CUSTOMER, SHOP_MANAGER, ADMIN]
    UserStatus:
      type: string
      enum: [VALID, SUSPENDED, WAITING]
    ShopStatus:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED]

    # --- Auth ---
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        address:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time

    # --- Shops ---
...
    Shop:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        manager:
          $ref: '#/components/schemas/UserProfile'
        category:
          $ref: '#/components/schemas/Category'
        note:
          type: number
          minimum: 0
          maximum: 5
          description: Note moyenne calculée à partir des avis
        hours:
          type: string
        image:
          type: string
        status:
          $ref: '#/components/schemas/ShopStatus'
        createdAt:
          type: string
          format: date-time
...
    ArticleInput:
      type: object
      required: [title, content, price]
      properties:
        title:
          type: string
        content:
          type: string
        price:
          type: number
    Article:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        price:
          type: number
        createdAt:
          type: string
          format: date-time